ZP-1284 Stickynote backend initial code upload.
This is running here against a BlackBerry DTEK66 and Priv, but may be a bit
rough around the edges and/or contain (plenty of) bugs!  Comments and
other's experiences with this code are welcome; please have at it!  Make
sure you read the REQUIREMENTS file so you understand what has to be set up.

Released under the Affero GNU General Public License (AGPL) version 3.