ZP-1391 Allow disabling the GSSAPI authenticator. Released under the Affero GNU General Public License (AGPL) version 3.
The php-imap module does not support proper fallback from GSSAPI to
other authentication methods (such as PLAIN). This causes the IMAP
backend to be non-functional when connecting to a GSSAPI-enabled IMAP
server (users will be unable to log in).
Since there is no fix available for the php-imap bug, this change
introduces a IMAP_DISABLE_AUTHENTICATOR setting that can be used to
manually disable GSSAPI support in the imap module.

Cf.: https://github.com/barbushin/php-imap/issues/19