Restore CountChanges timer to SyncCollections.
This was removed from z-push several years ago with the switch to changesink and the kopano backend. It introduced a bug in the IMAP backend where changes were not always detected with changesink and new messages are not pushed to mobile until more than one new message arrives. Reverting to the timer method may be too resource intensive on large deployments but it was the best workaround short of new IMAP code to better utilize changesink. There was some discussion in the now defunct z-hub jira.